Move encryption to encryption.py
